Have mas-field code in merch.js
