+++ libs/blocks/merch/merch.js
@@ -3,6 +3,8 @@ import {
   shouldAllowKrTrial, getCountry, getValidatedMasLibsUrl,
 } from '../../utils/utils.js';
 import { replaceKey } from '../../features/placeholders.js';
+import { decorateButtons, getBlockSize } from '../../utils/decorate.js';
+import { localizePreviewLinks, decorateContentLinks } from './autoblock.js';
 
 // MAS Component Names
 export const COMMERCE_LIBRARY = 'commerce';
@@ -1649,37 +1651,6 @@ export function getOptions(el) {
   return options;
 }
 
-export default async function init(el) {
-  if (!el?.classList?.contains('merch')) return undefined;
-  const url = new URL(el.href);
-  // Inline fragment field links (mas.adobe.com/studio.html#...&field=...) are routed here
-  // instead of merch-card-autoblock (see decorateAutoBlock in utils.js) so a field render
-  // never pulls in merch-card. Kept lazy so plain price/CTA links don't pay for it either.
-  if (url.hash.includes('field=')) {
-    const { default: initMasField } = await import('./mas-field.js');
-    return initMasField(el);
-  }
-  const { searchParams } = url;
-  const isCta = searchParams.get('type') === 'checkoutUrl';
-  const merch = await (isCta ? buildCta : buildPrice)(el, searchParams);
-  const service = await initService();
-  log = service.Log.module('merch');
-  if (merch) {
-    log.debug('Rendering:', { options: { ...merch.dataset }, merch, el });
-    // Rebuilt merch href keeps only the OSI; stash the original for
-    // the "Edit OSI" badge.
-    if (getConfig()?.mep?.preview) {
-      const { mepMasStudioUrls } = await import('./mas-mep-utils.js');
-      mepMasStudioUrls.set(merch, el.href);
-      merch.dataset.masBlock = 'ost';
-    }
-    el.replaceWith(merch);
-    return merch;
-  }
-  log.warn('Failed to get context:', { el });
-  return null;
-}
-
 const MAS_FRAGMENT_API = 'https://www.adobe.com/mas/io/fragment';
 const MAS_FRAGMENT_API_KEY = 'wcms-commerce-ims-ro-user-milo';
 
@@ -1706,6 +1677,278 @@ export async function createFragmentErrorEl(uuid, label = 'Card', status = null)
   const idEl = createTag('p', { class: 'mas-frag-error-id' }, uuid || 'unknown');
   el.append(badgeEl, labelEl, idEl);
   return el;
+}
+
+// Renders a single inline fragment field (price / description / CTA) without loading
+// merch-card, so a field-only link (e.g. authored in a marquee) doesn't pull in
+// merch-card's own dependencies.
+
+const FIELD_TIMEOUT = 5000;
+const seenFragments = new Set();
+let fieldLog;
+
+const HEADING_SELECTOR = 'h1, h2, h3, h4, h5, h6';
+const BLOCK_CONTENT_SELECTOR = `${HEADING_SELECTOR}, p, div, ul, ol, table, blockquote, pre, figure, section, article, hr`;
+const INLINE_WRAPPER_SELECTOR = 'strong, em, span, b, i, u, small, mark';
+
+/**
+ * Upgrades plain commerce elements (missing `is` attribute) to their proper
+ * customized built-in equivalents so the commerce service resolves them.
+ * e.g. <a data-wcs-osi="..."> → <a is="checkout-link" data-wcs-osi="...">
+ */
+const COMMERCE_IS_BY_TAG = { a: 'checkout-link', button: 'checkout-button', span: 'inline-price' };
+function upgradeCommerceLinks(content) {
+  content.querySelectorAll('[data-wcs-osi]:not([is])').forEach((el) => {
+    const isValue = COMMERCE_IS_BY_TAG[el.tagName.toLowerCase()];
+    if (!isValue) return;
+    const upgraded = document.createElement(el.tagName.toLowerCase(), { is: isValue });
+    [...el.attributes].forEach(({ name, value }) => upgraded.setAttribute(name, value));
+    upgraded.innerHTML = el.innerHTML;
+    el.replaceWith(upgraded);
+  });
+}
+
+function getFieldTimeoutPromise() {
+  return new Promise((resolve) => {
+    setTimeout(() => resolve('timeout'), FIELD_TIMEOUT);
+  });
+}
+
+async function loadFieldDependencies() {
+  const servicePromise = initService();
+  const success = await Promise.race([servicePromise, getFieldTimeoutPromise()]);
+  if (!success) {
+    throw new Error('Failed to initialize mas commerce service');
+  }
+  const service = await servicePromise;
+  fieldLog = service.Log.module('mas-field');
+  await loadMasComponent(MAS_FIELD);
+}
+
+async function checkFieldReady(masField, fragment) {
+  if (isMasErrorEnv()) {
+    const uuid = fragment ?? masField.querySelector('aem-fragment')?.getAttribute('fragment');
+    if (masField.hasAttribute('failed')) {
+      createFragmentErrorEl(uuid, 'Field').then((el) => masField.insertAdjacentElement('beforebegin', el));
+    } else {
+      masField.addEventListener('aem:error', async (e) => {
+        masField.insertAdjacentElement('beforebegin', await createFragmentErrorEl(uuid, 'Field', e.detail?.status));
+      }, { once: true });
+    }
+  }
+
+  const readyPromise = masField.checkReady();
+  const success = await Promise.race([readyPromise, getFieldTimeoutPromise()]);
+  if (success === 'timeout') {
+    fieldLog.error(`${masField.tagName} did not initialize withing give timeout`);
+  } else if (!success) {
+    fieldLog.error(`${masField.tagName} failed to initialize`);
+  }
+}
+
+function hasOnlyTargetContent(parent, target) {
+  if (!parent || !target || target.parentElement !== parent) return false;
+  return [...parent.childNodes].every((node) => {
+    if (node === target) return true;
+    return node.nodeType === Node.TEXT_NODE && node.textContent.trim() === '';
+  });
+}
+
+function unwrapInlineWrappers(masField) {
+  let parent = masField.parentElement;
+  while (parent?.matches?.(INLINE_WRAPPER_SELECTOR)
+    && hasOnlyTargetContent(parent, masField)) {
+    parent.replaceWith(masField);
+    parent = masField.parentElement;
+  }
+}
+
+function normalizeBlockFieldWrappers(masField) {
+  const content = masField.querySelector(':scope > [data-role="mas-field-content"]');
+  if (!content?.querySelector(BLOCK_CONTENT_SELECTOR)) return;
+
+  unwrapInlineWrappers(masField);
+
+  const parent = masField.parentElement;
+  if (!parent || !hasOnlyTargetContent(parent, masField)) return;
+
+  if (parent.matches(HEADING_SELECTOR)) {
+    const innerHeading = [...content.children]
+      .find((child) => child.matches?.(HEADING_SELECTOR));
+    if (innerHeading) {
+      if (parent.id && !innerHeading.id) innerHeading.id = parent.id;
+      parent.classList.forEach((className) => innerHeading.classList.add(className));
+    }
+  }
+
+  if (parent.matches('p') || parent.matches(HEADING_SELECTOR)) {
+    parent.replaceWith(masField);
+  }
+}
+
+function copyMasFieldIdToParent(masField, name) {
+  if (masField.getAttribute(name)) {
+    masField.parentElement.setAttribute(`data-mas-field-${name}`, masField.getAttribute(name));
+  }
+}
+
+function preserveInlineCommerceContext(masField, content) {
+  const promotionCode = masField.getAttribute('data-promotion-code');
+  if (promotionCode) {
+    content.querySelectorAll('span[is="inline-price"]:not([data-promotion-code]), a[is="checkout-link"]:not([data-promotion-code]), button[is="checkout-button"]:not([data-promotion-code])')
+      .forEach((commerceEl) => commerceEl.setAttribute('data-promotion-code', promotionCode));
+  }
+  if (content.querySelector('span[is="inline-price"]')) {
+    loadStyle(`${getConfig().base}/blocks/merch/merch.css`);
+  }
+}
+
+/**
+ * Hoists a resolved inline CTA into the authored em/strong and runs decorateButtons.
+ * Deferred until every CTA mas-field in the container is hoisted, so a still-wrapped
+ * sibling isn't matched by 'em a'/'strong a' with the wrong parent (its content span).
+ */
+function decorateInlineCtas(masField, content) {
+  const container = masField.closest('p, div');
+
+  // The block this CTA belongs to (direct child of a section). Bounds the sibling
+  // lookup so a foreign block's button can't dictate this CTA's size.
+  let blockEl = container?.parentElement;
+  while (blockEl?.parentElement && !blockEl.parentElement.classList.contains('section')
+    && blockEl.parentElement !== document.body) {
+    blockEl = blockEl.parentElement;
+  }
+
+  // Only a sized sibling counts: mas-field self-styles CTAs with con-button but no size,
+  // and using an unsized one as reference would drop button-xl on every CTA.
+  const SIZE_CLASS = /^button-(s|m|l|xl)$/;
+  const siblingBtn = [...(blockEl?.querySelectorAll('.con-button') ?? [])]
+    .find((b) => !masField.contains(b) && [...b.classList].some((c) => SIZE_CLASS.test(c)));
+  let size;
+  let utilClasses = [];
+
+  if (siblingBtn) {
+    // Inherit the sibling's exact size and utility classes.
+    size = [...siblingBtn.classList].find((c) => SIZE_CLASS.test(c));
+    utilClasses = [...siblingBtn.classList].filter((c) => c.startsWith('button-') && c !== size);
+  } else {
+    // No decorated sibling yet — derive size from the block, mirroring its decorateButtons call.
+    const btnVariant = [...(blockEl?.classList ?? [])].find((c) => c.endsWith('-button'));
+    if (btnVariant) {
+      size = `button-${btnVariant.split('-')[0]}`;
+    } else if (blockEl?.classList.contains('hero-marquee')) {
+      // Mirror hero-marquee's extendButtonsClass; the util class keeps CTAs full-width on mobile.
+      size = 'button-xl';
+      utilClasses = ['button-justified-mobile'];
+    } else if (blockEl?.classList.contains('accordion') || blockEl?.classList.contains('media')) {
+      size = null;
+    } else {
+      const blockSize = getBlockSize(blockEl ?? container);
+      size = (blockSize === 'large' || blockSize === 'xlarge') ? 'button-xl' : 'button-l';
+    }
+  }
+  copyMasFieldIdToParent(masField, 'fragment-id');
+  copyMasFieldIdToParent(masField, 'variation-id');
+  preserveInlineCommerceContext(masField, content);
+  masField.replaceWith(...content.childNodes);
+
+  const pendingCTAs = container?.querySelectorAll('em > mas-field, strong > mas-field');
+  if (container && !pendingCTAs?.length) {
+    decorateButtons(container, size);
+    if (utilClasses.length) {
+      container.querySelectorAll('.con-button').forEach((b) => utilClasses.forEach((c) => b.classList.add(c)));
+    }
+  }
+}
+
+/**
+ * A headless mas-field CTA can fire 'mas:ready' after its block decorated (slow network),
+ * too late for decorateButtons. One document listener hoists + decorates it, no per-block wiring.
+ */
+let masReadyWatched = false;
+function watchMasFieldCtas() {
+  if (masReadyWatched) return;
+  masReadyWatched = true;
+  document.addEventListener('mas:ready', async ({ target: mf }) => {
+    if (mf?.tagName !== 'MAS-FIELD' || !mf.closest('em, strong')) return;
+    const content = mf.querySelector(':scope > [data-role="mas-field-content"]');
+    // Same gate createInline uses: an inline CTA anchor, not block-level content.
+    if (content?.querySelector('a') && !content.querySelector(BLOCK_CONTENT_SELECTOR)) {
+      // Upgrade to checkout-link before hoisting, else the late CTA never hydrates.
+      upgradeCommerceLinks(content);
+      await decorateContentLinks(content);
+      decorateInlineCtas(mf, content);
+    }
+  });
+}
+
+/** Replaces an inline fragment link with a mas-field wrapping an aem-fragment. */
+async function createInlineField(el, options) {
+  const aemFragment = createAemFragment(options, seenFragments);
+  const masField = createTag('mas-field', { field: options.field }, aemFragment);
+  if (getConfig()?.mep?.preview) {
+    const { mepMasStudioUrls } = await import('./mas-mep-utils.js');
+    mepMasStudioUrls.set(masField, el.href);
+    masField.dataset.masBlock = 'inline';
+  }
+  el.replaceWith(masField);
+  await checkFieldReady(masField, options.fragment);
+  normalizeBlockFieldWrappers(masField);
+  await localizePreviewLinks(masField);
+
+  const content = masField.querySelector(':scope > [data-role="mas-field-content"]');
+  if (!content) return;
+
+  // Upgrade any plain commerce elements (missing `is`) so the commerce service resolves
+  // them. Applies to both CTA (<a>) and price (<span>) fields.
+  upgradeCommerceLinks(content);
+
+  await decorateContentLinks(content);
+
+  // Inline CTAs: hoist the anchor into the authored em/strong and let decorateButtons style it.
+  if (content.querySelector('a') && !content.querySelector(BLOCK_CONTENT_SELECTOR)) {
+    decorateInlineCtas(masField, content);
+  }
+}
+
+export async function initMasField(el) {
+  watchMasFieldCtas();
+  let options = getOptions(el);
+  const { fragment } = options;
+  if (!fragment) return;
+  options = overrideOptions(fragment, options);
+  await loadFieldDependencies();
+  await createInlineField(el, options);
+}
+
+export default async function init(el) {
+  if (!el?.classList?.contains('merch')) return undefined;
+  const url = new URL(el.href);
+  // Inline fragment field links (mas.adobe.com/studio.html#...&field=...) are routed here
+  // instead of merch-card-autoblock (see decorateAutoBlock in utils.js) so a field render
+  // never pulls in merch-card.
+  if (url.hash.includes('field=')) {
+    return initMasField(el);
+  }
+  const { searchParams } = url;
+  const isCta = searchParams.get('type') === 'checkoutUrl';
+  const merch = await (isCta ? buildCta : buildPrice)(el, searchParams);
+  const service = await initService();
+  log = service.Log.module('merch');
+  if (merch) {
+    log.debug('Rendering:', { options: { ...merch.dataset }, merch, el });
+    // Rebuilt merch href keeps only the OSI; stash the original for
+    // the "Edit OSI" badge.
+    if (getConfig()?.mep?.preview) {
+      const { mepMasStudioUrls } = await import('./mas-mep-utils.js');
+      mepMasStudioUrls.set(merch, el.href);
+      merch.dataset.masBlock = 'ost';
+    }
+    el.replaceWith(merch);
+    return merch;
+  }
+  log.warn('Failed to get context:', { el });
+  return null;
 }
 
 window.addEventListener('hashchange', updateModalState);
